requirements.txt
pandas
numpy
matplotlib
seaborn
statsmodel
sklearn
joblib
PyYAML-6.0.2